Optimize components

+++ backend/controllers/cartController.js
@@ -5,6 +5,10 @@ const addToCart = async(req,res) => {
     try {
         const {productId, size} = req.body;
         const userId = req.user._id;
+        
+        if (!productId) return res.status(404).json({error: "ProductId not found"});
+        if (!size) return res.status(404).json({error: "size not found"});
+        if (!userId) return res.status(404).json({error: "userId not found"});
 
         const existsCart = await cartModel.findOne({ productId, userId, size });
         
@@ -41,10 +45,12 @@ const addToCart = async(req,res) => {
 const getCartItems = async(req,res) => {
     try {
         const userId = req.user._id;
-        
+        if (!userId) return res.status(404).json({error: "userId not found"});
+
         const cartItems = await cartModel.find({userId})
         .sort({createdAt: -1})
         .populate({path: 'productId', select: "_id category subCategory name images color discount price vendorId brandName"})
+
         if(!cartItems) return res.status(404).json({error: "Cart items not found"});
         
         res.status(200).json(cartItems);
@@ -59,6 +65,8 @@ const getCartItems = async(req,res) => {
 const updateQuantity = async(req,res) => {
     try {
         const {cartId, quantity} = req.body;
+        if (!cartId) return res.status(404).json({error: "cartId not found"});
+        if (!quantity) return res.status(404).json({error: "quantity not found"});
 
         const updateCartQty = await cartModel.findById(cartId);
         if (!updateCartQty) {
@@ -79,6 +87,7 @@ const updateQuantity = async(req,res) => {
 const DeleteCart = async(req,res) => {
     try {
         const {cartId} = req.body;
+        if (!cartId) return res.status(404).json({error: "cartId not found"});
 
         const deleteCart = await cartModel.findByIdAndDelete(cartId);
         if (!deleteCart) {
@@ -91,9 +100,11 @@ const DeleteCart = async(req,res) => {
     }
 }
 
+
 const DeleteUserCarts = async(req,res) => {
     try {
         const userId = req.user._id;
+        if (!userId) return res.status(404).json({error: "userId not found"});
 
         const deleteCarts = await cartModel.deleteMany({ userId: userId });
